requirements.txt
requests>=2.32.3
beautifulsoup4>=4.12.3
selenium>=4.28.1
python-dotenv>=1.0.1
